Revert "Move blackbox about simulating with VERILATOR_THREADS to simulation"
This reverts commit c91b20efcb33be4b144c88a2e1cf5aa8ca8a1b02.

It is _actually_ required to *BUILD* the design using this makefile
variable during the building process rather than the simulation
process.

As for the exact requirements to make the simulation multithreaded, I
am not sure. I know that you *MUST* provide the VERILATOR_THREADS
variable during the build process, but am unsure if it must be
provided again during simulation time.